Moved HasCase class into MatchCI, leaving StringLike unchanged
